requirements.txt
requests
schedule
python-dotenv
playwright
